Move learn style code to corret position
